fix(popup-menu-web): refactor and fix render glitch
